Rename Chunk stride test to match what it actually verifies
Row_stride_is_derived_from_the_types_not_from_a_literal could not
distinguish sizeof(long) from Unsafe.SizeOf<Entity>() since both are
8 bytes today, so it would stay green even if the fix were reverted.
Renamed it to Multi_row_addressing_stays_consistent_across_accessors
and added Layout_reserves_room_for_every_row_at_the_types_stride,
which asserts on ChunkLayout offsets directly and does fail when
EntitySize is perturbed (verified manually, then reverted).

Also aligned GetRef<T>'s row multiplication to the (nint)-first form
used by EntityAt/TagAt, and gave TagAt a rationale doc comment.

+++ src/UvEcs/Chunk.cs
@@ -46,6 +46,7 @@ public sealed unsafe class Chunk
         return ref Unsafe.AsRef<Entity>((void*)(_data + Layout.EntityOffset + (nint)row * Unsafe.SizeOf<Entity>()));
     }
 
+    /// <remarks>Шаг строки берётся из типа, а не из литерала — см. <see cref="EntityAt"/>.</remarks>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public ref TagMask TagAt(int row)
     {
@@ -70,7 +71,7 @@ public sealed unsafe class Chunk
     public ref T GetRef<T>(int row) where T : unmanaged, IComponent
     {
         if ((uint)row >= (uint)Count) throw new ArgumentOutOfRangeException(nameof(row));
-        return ref Unsafe.AsRef<T>((void*)(_data + Layout.ColumnOffsets[ColumnOrThrow<T>()] + row * Unsafe.SizeOf<T>()));
+        return ref Unsafe.AsRef<T>((void*)(_data + Layout.ColumnOffsets[ColumnOrThrow<T>()] + (nint)row * Unsafe.SizeOf<T>()));
     }
 
     /// <summary>
